+++ .github/scripts/stale-eng-issues.test.js
@@ -4,12 +4,10 @@ const test = require('node:test');
 const assert = require('node:assert');
 
 const run = require('./stale-eng-issues.js');
-// Pull constants from the script so a future policy change (e.g. STALE_DAYS 365 -> 730) surfaces
-// in the boundary tests instead of silently passing because the test hardcoded the old value.
+// Pull constants from the script so the boundary tests keep exercising the real boundary if a
+// future policy change (e.g. STALE_DAYS 365 -> 730) moves it.
 const { STALE_DAYS, CLOSE_DAYS, ELIGIBLE_LABEL } = run;
-
-const DAY_MS = 24 * 60 * 60 * 1000;
-const daysAgoIso = (days) => new Date(Date.now() - days * DAY_MS).toISOString();
+const { DAY_MS, daysAgoIso, makeStaleLabelEvent, makeContext, makeCore, makeGithub } = require('./stale-test-helpers.js');
 
 // Eng-initiated issues qualify by carrying the `~engineering-initiated` label, so the default
 // issue includes it. Tests that exercise the ineligible path drop it explicitly.
@@ -26,112 +24,6 @@ function makeIssue(overrides = {}) {
   };
 }
 
-function makeStaleLabelEvent({ daysAgo = 100, at } = {}) {
-  const created_at = at != null ? new Date(at).toISOString() : daysAgoIso(daysAgo);
-  return { event: 'labeled', label: { name: 'stale' }, created_at };
-}
-
-function makeContext() {
-  return { repo: { owner: 'o', repo: 'r' } };
-}
-
-function makeCore() {
-  const infos = [];
-  const warnings = [];
-  const summaryCalls = [];
-  const summary = new Proxy(
-    {},
-    {
-      get(_target, prop) {
-        if (prop === 'write') {
-          return async () => {
-            summaryCalls.push({ method: 'write' });
-          };
-        }
-        return (...args) => {
-          summaryCalls.push({ method: String(prop), args });
-          return summary;
-        };
-      },
-    },
-  );
-  return {
-    info: (msg) => infos.push(msg),
-    warning: (msg) => warnings.push(msg),
-    summary,
-    _captured: { infos, warnings, summaryCalls },
-  };
-}
-
-function makeGithub({ issuesByPage = [], eventsByIssue = {}, failOn = {} } = {}) {
-  const createCommentCalls = [];
-  const addLabelsCalls = [];
-  const removeLabelCalls = [];
-  const updateCalls = [];
-  const listEventsCalls = [];
-
-  const counters = {};
-  const shouldFail = (op) => {
-    const cfg = failOn[op];
-    if (!cfg) return null;
-    counters[op] = (counters[op] || 0) + 1;
-    if (cfg === 'always') return new Error(`${op} simulated failure`);
-    if (typeof cfg === 'number' && counters[op] <= cfg) return new Error(`${op} simulated failure`);
-    if (typeof cfg === 'object' && cfg.status && counters[op] === 1) {
-      const err = new Error(`${op} simulated failure status ${cfg.status}`);
-      err.status = cfg.status;
-      return err;
-    }
-    return null;
-  };
-
-  const paginate = async (endpoint, params) => {
-    if (endpoint === 'listEvents-sentinel') {
-      listEventsCalls.push(params);
-      const err = shouldFail('listEvents');
-      if (err) throw err;
-      return eventsByIssue[params.issue_number] || [];
-    }
-    return [];
-  };
-  paginate.iterator = async function* iterator(endpoint) {
-    if (endpoint === 'listForRepo-sentinel') {
-      for (const page of issuesByPage) yield { data: page };
-    }
-  };
-
-  return {
-    paginate,
-    rest: {
-      issues: {
-        listForRepo: 'listForRepo-sentinel',
-        listEvents: 'listEvents-sentinel',
-        createComment: async (params) => {
-          const err = shouldFail('createComment');
-          if (err) throw err;
-          createCommentCalls.push(params);
-        },
-        addLabels: async (params) => {
-          const err = shouldFail('addLabels');
-          if (err) throw err;
-          addLabelsCalls.push(params);
-        },
-        removeLabel: async (params) => {
-          const err = shouldFail('removeLabel');
-          if (err) throw err;
-          removeLabelCalls.push(params);
-        },
-        update: async (params) => {
-          const err = shouldFail('update');
-          if (err) throw err;
-          updateCalls.push(params);
-        },
-      },
-    },
-    _captured: { createCommentCalls, addLabelsCalls, removeLabelCalls, updateCalls, listEventsCalls },
-  };
-}
-
 async function runWith({ issues, issuesByPage, dryRun = false, maxOps = 1000, eventsByIssue = {}, failOn = {} } = {}) {
   process.env.DRY_RUN = dryRun ? 'true' : 'false';
   process.env.MAX_OPERATIONS = String(maxOps);
@@ -143,12 +35,19 @@ async function runWith({ issues, issuesByPage, dryRun = false, maxOps = 1000, ev
   return { github, core, result };
 }
 
+// Core engine behavior (caps, error paths, pagination, un-stale epsilon, etc.) is covered through
+// the Fleetie wrapper in stale-fleetie-issues.test.js. These tests pin what the eng wrapper
+// configures differently: label-based eligibility, no exempt labels, the 1-year threshold, and the
+// eng-specific message wording.
+
 test('marks an eng-initiated issue idle >1y as stale and @-mentions the author', async () => {
   const { github, result } = await runWith({
     issues: [makeIssue({ number: 1, user: { login: 'getvictor' }, updated_at: daysAgoIso(STALE_DAYS + 70) })],
   });
   assert.strictEqual(github._captured.createCommentCalls.length, 1);
-  assert.match(github._captured.createCommentCalls[0].body, /^@getvictor /, 'stale comment @-mentions the author');
+  const staleComment = github._captured.createCommentCalls[0].body;
+  assert.match(staleComment, /^@getvictor /, 'stale comment @-mentions the author');
+  assert.ok(staleComment.includes('365 days'), 'stale comment uses the eng wording, not the Fleetie template');
   assert.strictEqual(github._captured.addLabelsCalls.length, 1);
   assert.deepStrictEqual(github._captured.addLabelsCalls[0].labels, ['stale']);
   assert.strictEqual(result.staled.length, 1);
@@ -179,7 +78,7 @@ test('does not exempt bug-labeled eng issues (unlike the Fleetie closer)', async
 });
 
 test('closes a stale-labeled eng issue idle >14d with no activity after labeling', async () => {
-  const labeledAt = Date.now() - 20 * DAY_MS;
+  const labeledAt = Date.now() - (CLOSE_DAYS + 6) * DAY_MS;
   const issue = makeIssue({
     number: 2,
     labels: [{ name: ELIGIBLE_LABEL }, { name: 'stale' }],
@@ -189,6 +88,7 @@ test('closes a stale-labeled eng issue idle >14d with no activity after labeling
     issues: [issue],
     eventsByIssue: { 2: [makeStaleLabelEvent({ at: labeledAt })] },
   });
+  assert.strictEqual(github._captured.createCommentCalls.length, 1, 'close comment posted');
   assert.strictEqual(github._captured.updateCalls.length, 1);
   assert.strictEqual(github._captured.updateCalls[0].state, 'closed');
   assert.strictEqual(github._captured.updateCalls[0].state_reason, 'not_planned');
@@ -207,6 +107,8 @@ test('un-stales an eng issue with activity after labeling', async () => {
   });
   assert.strictEqual(github._captured.removeLabelCalls.length, 1);
   assert.strictEqual(github._captured.removeLabelCalls[0].name, 'stale');
+  assert.strictEqual(github._captured.createCommentCalls.length, 0, 'un-stale writes no comment');
+  assert.strictEqual(github._captured.updateCalls.length, 0, 'un-stale does not close');
   assert.strictEqual(result.unstaled.length, 1);
 });
 
@@ -216,24 +118,6 @@ test('skips eng issue idle between 14 and 365 days as "not stale yet"', async ()
   });
   assert.strictEqual(result.skippedNotStaleYet, 1);
   assert.strictEqual(result.staled.length, 0);
-});
-
-test('excludes pull requests', async () => {
-  const { github } = await runWith({
-    issues: [makeIssue({ updated_at: daysAgoIso(STALE_DAYS + 70), pull_request: { url: 'x' } })],
-  });
-  assert.strictEqual(github._captured.createCommentCalls.length, 0);
-});
-
-test('dry-run never writes', async () => {
-  const { github, result } = await runWith({
-    issues: [makeIssue({ updated_at: daysAgoIso(STALE_DAYS + 70) })],
-    dryRun: true,
-  });
-  assert.strictEqual(github._captured.createCommentCalls.length, 0);
-  assert.strictEqual(github._captured.addLabelsCalls.length, 0);
-  assert.strictEqual(result.dryRun, true);
-  assert.strictEqual(result.staled.length, 1);
 });
 
 test('stale-phase boundary: issue just past STALE_DAYS is staled', async () => {
