+++ .github/workflows/test-stale-fleetie-issue-scripts.yml
@@ -15,6 +15,7 @@ on:
       - '.github/scripts/build-fleetie-handles.js'
       - '.github/scripts/build-fleetie-handles.test.js'
       - '.github/scripts/stale-issues-core.js'
+      - '.github/scripts/stale-test-helpers.js'
       - '.github/scripts/stale-eng-issues.js'
       - '.github/scripts/stale-eng-issues.test.js'
       - '.github/scripts/stale-fleetie-issues.js'
@@ -27,6 +28,7 @@ on:
       - '.github/scripts/build-fleetie-handles.js'
       - '.github/scripts/build-fleetie-handles.test.js'
       - '.github/scripts/stale-issues-core.js'
+      - '.github/scripts/stale-test-helpers.js'
       - '.github/scripts/stale-eng-issues.js'
       - '.github/scripts/stale-eng-issues.test.js'
       - '.github/scripts/stale-fleetie-issues.js'
